+++ sdk/python/tests/unit/test_feature_views.py
@@ -1,22 +1,16 @@
-import copy
 from datetime import timedelta
 
 import pytest
 from typeguard import TypeCheckError
 
-from feast.aggregation import Aggregation
 from feast.batch_feature_view import BatchFeatureView
 from feast.data_format import AvroFormat
-from feast.data_source import KafkaSource, PushSource
+from feast.data_source import KafkaSource
 from feast.entity import Entity
 from feast.feature_view import FeatureView
 from feast.field import Field
 from feast.infra.offline_stores.file_source import FileSource
-from feast.protos.feast.core.StreamFeatureView_pb2 import (
-    StreamFeatureView as StreamFeatureViewProto,
-)
 from feast.protos.feast.types.Value_pb2 import ValueType
-from feast.stream_feature_view import StreamFeatureView, stream_feature_view
 from feast.types import Float32
 
 
@@ -65,167 +59,8 @@ def test_create_batch_feature_view():
         )
 
 
-def test_create_stream_feature_view():
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-    StreamFeatureView(
-        name="test kafka stream feature view",
-        entities=[],
-        ttl=timedelta(days=30),
-        source=stream_source,
-        aggregations=[],
-    )
-
-    push_source = PushSource(
-        name="push source", batch_source=FileSource(path="some path")
-    )
-    StreamFeatureView(
-        name="test push source feature view",
-        entities=[],
-        ttl=timedelta(days=30),
-        source=push_source,
-        aggregations=[],
-    )
-
-    with pytest.raises(TypeError):
-        StreamFeatureView(
-            name="test batch feature view",
-            entities=[],
-            ttl=timedelta(days=30),
-            aggregations=[],
-        )
-
-    with pytest.raises(ValueError):
-        StreamFeatureView(
-            name="test batch feature view",
-            entities=[],
-            ttl=timedelta(days=30),
-            source=FileSource(path="some path"),
-            aggregations=[],
-        )
-
-
 def simple_udf(x: int):
     return x + 3
-
-
-def test_stream_feature_view_serialization():
-    entity = Entity(name="driver_entity", join_keys=["test_key"])
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-
-    sfv = StreamFeatureView(
-        name="test kafka stream feature view",
-        entities=[entity],
-        ttl=timedelta(days=30),
-        owner="test@example.com",
-        online=True,
-        schema=[Field(name="dummy_field", dtype=Float32)],
-        description="desc",
-        aggregations=[
-            Aggregation(
-                column="dummy_field",
-                function="max",
-                time_window=timedelta(days=1),
-            )
-        ],
-        timestamp_field="event_timestamp",
-        mode="spark",
-        source=stream_source,
-        udf=simple_udf,
-        tags={},
-    )
-
-    sfv_proto = sfv.to_proto()
-
-    new_sfv = StreamFeatureView.from_proto(sfv_proto=sfv_proto)
-    assert new_sfv == sfv
-
-
-def test_stream_feature_view_udfs():
-    entity = Entity(name="driver_entity", join_keys=["test_key"])
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-
-    @stream_feature_view(
-        entities=[entity],
-        ttl=timedelta(days=30),
-        owner="test@example.com",
-        online=True,
-        schema=[Field(name="dummy_field", dtype=Float32)],
-        description="desc",
-        aggregations=[
-            Aggregation(
-                column="dummy_field",
-                function="max",
-                time_window=timedelta(days=1),
-            )
-        ],
-        timestamp_field="event_timestamp",
-        source=stream_source,
-    )
-    def pandas_udf(pandas_df):
-        import pandas as pd
-
-        assert type(pandas_df) == pd.DataFrame
-        df = pandas_df.transform(lambda x: x + 10, axis=1)
-        return df
-
-    import pandas as pd
-
-    df = pd.DataFrame({"A": [1, 2, 3], "B": [10, 20, 30]})
-    sfv = pandas_udf
-    sfv_proto = sfv.to_proto()
-    new_sfv = StreamFeatureView.from_proto(sfv_proto)
-    new_df = new_sfv.udf(df)
-
-    expected_df = pd.DataFrame({"A": [11, 12, 13], "B": [20, 30, 40]})
-
-    assert new_df.equals(expected_df)
-
-
-def test_stream_feature_view_initialization_with_optional_fields_omitted():
-    entity = Entity(name="driver_entity", join_keys=["test_key"])
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-
-    sfv = StreamFeatureView(
-        name="test kafka stream feature view",
-        entities=[entity],
-        schema=[],
-        description="desc",
-        timestamp_field="event_timestamp",
-        source=stream_source,
-        tags={},
-    )
-    sfv_proto = sfv.to_proto()
-
-    new_sfv = StreamFeatureView.from_proto(sfv_proto=sfv_proto)
-    assert new_sfv == sfv
 
 
 def test_hash():
@@ -282,41 +117,3 @@ def test_hash():
 def test_field_types():
     with pytest.raises(TypeCheckError):
         Field(name="name", dtype=ValueType.INT32)
-
-
-def test_stream_feature_view_proto_type():
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-    sfv = StreamFeatureView(
-        name="test stream featureview proto class",
-        entities=[],
-        ttl=timedelta(days=30),
-        source=stream_source,
-        aggregations=[],
-    )
-    assert sfv.proto_class is StreamFeatureViewProto
-
-
-def test_stream_feature_view_copy():
-    stream_source = KafkaSource(
-        name="kafka",
-        timestamp_field="event_timestamp",
-        kafka_bootstrap_servers="",
-        message_format=AvroFormat(""),
-        topic="topic",
-        batch_source=FileSource(path="some path"),
-    )
-    sfv = StreamFeatureView(
-        name="test stream featureview proto class",
-        entities=[],
-        ttl=timedelta(days=30),
-        source=stream_source,
-        aggregations=[],
-    )
-    assert sfv == copy.copy(sfv)
